requirements.txt
# nnPU dependencies (from nnPU/env.yaml)
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
scipy==1.10.1
scikit-learn==1.2.2
torchvision==0.15.2
ipykernel==6.23.1
ipywidgets
transformers==4.31.0
datasets==2.15.0
sentence-transformers==2.2.2
pillow==8.4.0
pkbar==0.5
pytest==7.3.2
altair==5.0.1
vl-convert-python==0.11.2
cairosvg==2.7.0
huggingface_hub==0.25.2

# DRPU dependencies
cvxopt==1.3.2

# PULS dependencies
pydantic==2.8.2